Catch errors (synchronously) thrown by `onSubmit`
The presence of `onSubmitFailure` makes it look like it might be ok to let `onSubmit` throw an error (as well as rejecting the promise it returns). If `onSubmit` does throw an error, the form itself could enter an invalid state and the uncaught error message would be mixed up with uniforms implementation, so it would be hard for the end user to debug.